% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{incrementality_func}
\alias{incrementality_func}
\title{Statistic for the bootstrap (legacy interface)}
\usage{
incrementality_func(datadiff, indices)
}
\arguments{
\item{datadiff}{Numeric vector of differences.}

\item{indices}{Resampled indices.}
}
\value{
Mean and estimated variance of the mean.
}
\description{
Statistic for the bootstrap (legacy interface)
}
